wip lane moving
• path finding only if outgoing connection allows
• for runtime store speed in m/s
• remove restriction of 1000 vehicles in sim

+++ CitySimulator/src/core/include/core/simulation/movement/lane_agnostic_movement.h
@@ -11,7 +11,7 @@ namespace tjs::core::simulation {
 	struct LaneRuntime {
 		Lane* static_lane = nullptr;
 		double length;
-		int max_speed;
+		float max_speed;
 		std::vector<std::size_t> idx;
 	};
